refactor: move role-name into 'models' folder
