+++ src/controllers/google.ts
@@ -8,7 +8,7 @@ import {
   getIdToken
 } from "../services/google-auth.service";
 import { User } from "../models/interface/user";
-import { RoleName } from "../services/authz.service/role-name";
+import { RoleName } from "../models/role-name";
 import { handleJsonData, handleErrors } from "./helpers";
 
 async function signInUpGoogle(payload: any) {
