+++ src/models/interface/user.ts
@@ -1,11 +1,11 @@
-import { RoleName } from '../../services/authz.service/role-name';
+import { RoleName } from "../role-name";
 
 export interface User {
-    id?: any;
-    name: string;
-    email: string;
-    roleName: RoleName;
-    password?: string;
-    facebookId?: string;
-    googleId?: string;
+  id?: any;
+  name: string;
+  email: string;
+  roleName: RoleName;
+  password?: string;
+  facebookId?: string;
+  googleId?: string;
 }
