+++ src/services/authz.service/rule.ts
@@ -1,7 +1,7 @@
-import { RoleName } from './role-name';
-import { HttpMethod } from './http-method';
+import { RoleName } from "../../models/role-name";
+import { HttpMethod } from "./http-method";
 
 export interface Rule {
-    httpMethod: HttpMethod,
-    roleNames: RoleName[]
+  httpMethod: HttpMethod;
+  roleNames: RoleName[];
 }
